requirements.txt
numpy
pandas
scikit-learn
python-dotenv
psycopg2-binary
supabase
google-generativeai
fastapi
uvicorn[standard]
pydantic
qrcode

